LoggingSolvers: updates from upstream
Updates from upstream KLEE to make the SMTLIB logging solver available in S2E.

One of the changes from upstream was the replacement of the PCLoggingSolver
with the KQueryLoggingSolver. The functionality is the same, except the new
logging solver uses a .kquery file extension rather than a .pc file extension.

I also had to bring across some updates for the ConstantExpr class, otherwise
ExprSMTLIBPrinter::printConstant would cause the assert in
ConstantExpr::getZExtValue to fail when calling isTrue (due to a mismatch in
bit widths). The fix for this (taken from upstream) is to check that expression
is a Bool type before calling getZExtValue.